PS Coding Test
a coding challenge I had to do for a job interview. I was to take a JSON file, filter out those that needed to have MFA enabled. Then compare the file to those in a CSV. Then output a CSV with those in the "engineering department" that didn't have MFA. If I were doing this in the real world I would pull from the service a response with just the info I needed. IE those in department A without MFA. Then pipe that to the service with a flag to enable MFA on those accounts.